[SC-96] style : 유저 게시글 반환, Post -> User로 이동
