+++ src/main/java/inu/codin/codin/domain/post/service/PostService.java
@@ -116,15 +116,6 @@ public class PostService {
 
     }
 
-
-    //해당 유저가 작성한 모든 글 반환 :: 게시글 내용 + 댓글+대댓글의 수 + 좋아요,스크랩 count 수 반환
-    public PostPageResponse getAllUserPosts(int pageNumber) {
-        ObjectId userId = SecurityUtils.getCurrentUserId();
-        PageRequest pageRequest = PageRequest.of(pageNumber, 20, Sort.by("createdAt").descending());
-        Page<PostEntity> page = postRepository.findAllByUserIdOrderByCreatedAt(userId, pageRequest);
-        return PostPageResponse.of(getPostListResponseDtos(page.getContent()), page.getTotalPages()-1, page.hasNext()? page.getPageable().getPageNumber() + 1 : -1);
-    }
-
     public List<PostListResponseDto> getPostListResponseDtos(List<PostEntity> posts) {
         return posts.stream()
                 .sorted(Comparator.comparing(PostEntity::getCreatedAt).reversed())
